api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
tensorflow>=2.16.0,<2.20.0
numpy>=1.24.0,<2.0.0
pillow>=10.0.0
opencv-python>=4.8.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
loguru>=0.7.0 